Add a smoke test for `AppMenu`
